backend/requirements.txt
boto3
black
python-dotenv
flask_restful
jwt
pyjwt[jwks_client]